Tech Challenge/Fase 5/requirements.txt
flask==3.1.1
iterative_stratification==0.1.9
Jinja2==3.1.6
Markdown==3.7
numpy==2.2.6
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.3.0
PyYAML==6.0.2
streamlit==1.46.1
ultralytics==8.3.161
